rewrite shell_exec as a plain old function
